[fix][subscribers] Regenerate metro status dict
**Why this change was necessary**
The metro status dictionary was not being regenerated correctly,
leading to type errors. There was also an incorrect comparison being
made in the logical conditions that would trigger a status update
being sent and saved.

**What this change does**
Refactors globals for default status messages and metro
colours. Refactors metro status dict and fixes conditional comparison
for triggering a status being posted in the server and updated in the dict.

**Any side-effects?**
None.

**Additional context/notes/links**
Note that this could not be adequately tested as metro outages are
infrequent.

Resolves: #221