[feat][core] Create test directory
**Why this change was necessary**
We'll have to write a test for each bot function, so tests must be
organized and in a separate directory.

**What this change does**
Creates the beginnings of a `tests` dir structure for future
tests. Contributing documentation and the README are also updated to
reflect this change.

**Any side-effects?**
A new dependency `pytest` was added, which will necessitate each
contributor to run `poetry install` on their local clones.

**Additional context/notes/links**

Resolves #250