Input emoji for editable content. Can return user
**Why this change was necessary**
Before, "editable_content=True" was only used for deletions, and the emoji was always the trash emoji. However, since the function only "gets out" of paginator, it can be useful for other reasons, for example to select an option to edit from a list.
Returning the user who clicked the button can also be useful, for example if they are the only one who will be able to edit.

**What this change does**
It is now possible to select the editable_content_emoji. The trash emoji is still the default one for compatibility with functions already using paginator.
It is now possible to return the user who clicked the button. By default, this isn't activated.
Renamed to "self.delete" to "self.edit_mode". Since this is used outside paginator to see when editable content is selected ("while p.paginator"), this is also changed in every concerned file (quotes.py and reminder.py)

**Any side-effects?**
There could be side-effects if I missed a change from "p.delete" to "p.edit_mode", however I don't think I did.