Add Custom Reactions Function
**Why this change was necessary**
This change is partly for the Reduce reliance on Mosebot project, but it also brings several improvements.
All moderators can now add custom reacts through an easy-to-use assistant, and choose options more easily and it is now easier to edit custom reacts.
Users (non-mods) can also propose custom reacts, which moderators can approve or reject.

**What this change does**
Creates a new customreactions cog. Adds a CustomReactions table in the database. Adds a new customreactions command. This command creates a wizard with three choices. Add (if mod)/Propose (if non-mod) a new custom reactions, see the list of custom reactions, see the list of proposals.
When adding/proposing a custom reactions, you must input a prompt, a response, and you can then choose if you want to select certain options:
-DeletePrompt, which deletes the message that calls the custom reaction
-Anywhere, which activates the custom reaction if the prompt is anywhere in a message
-DM, which reacts in the DMs of the user who calls the reaction instead of the channel
When listing the custom reactions/proposals, any user can see more info about a custom reaction (see the options, who created it, etc.). From there, mods can edit the prompt/response or the options, delete it, or, if it's a proposal, approve or reject it.
Adds an on_message event that, each time a message is sent, checks if the message is a prompt. The code is made so that the reaction_list isn't recreated each time, and only the messages that might be a prompt are then looked at more closely with the options etc.
If many custom reacts exist for that message, a random one is chosen.

**For the future**
Eventually, we could add "keyword" like %user%, which Mose had but was never used in the server