Remove hardcoded custom reactions
**Why this change was necessary**
With the new custom reaction command, we can move all of these two the same place, and then edit them from the server. For the ones that were commands, you can just add the prefix itself in the prompt (i.e. create a custom react with prompt "?gohere" for example).

**What this change does**
Removes the MARTY_RESPONSES dictionnary and the on_message event that were in main.py. Removes most of the memes command that can be transformed in custom reactions. I only let ?cheep there in case someone wants to test and doesnt have access to the database for example.