Add Reactions table, remove Score from Members
**Why this change was necessary**
The Reactions table is to get better stats on reactions. Score is now computed from Reactions and thus not needed anymore in Members. DisplayName is renamed Name to better reflect how it is now used (See below)

**What this change does**
Adds a Reactions table that, for every reaction, will keep the ID of the person that reacted (ReacterID),the ID of the person that wrote the message (ReacteeID), the name of the reaction (this is the result of str(emoji), that is, either the emoji directly for unicode directly, or <:emojiname:emojiid> for custom emoji), and the ID of the message. The ReacterID and ReacteeID reference IDs in the Members table. Each row is unique since a certain user can only react a certain emoji on a certain message once.

Since score is now calculated directly from that table, there is no need to cache score in the Members table anymore. DisplayName is renamed Name to avoid confusion since it will now use the username#descriminator instead (or if that is unavaible, the ID). The reference to DisplayName in currency.py is changed to Name.