Checks improvements, refuse attachments
msg_check now sends a message saying only urls can be used when a user send attachments in their prompt/response. It also checks that the message was sent in the same channel where custom reacts is used

Create get_msg_check, get_reaction_check and get_number_check. These three accepts a user as input, and will return a check that also checks that the message/reaction was sent by that user (this is useful if the user variable isn't "accessible" by the check). reaction_check also accepts a moderators boolean as input, which checks that the user reacting is a moderator.
number_check now accepts a range as input, and will check if the number is inside that check
(There were a couple "manual" checks, where for example there was a loop with a `wait_for` for a message, and then when `wait_for` returned a message, it checked if it was from a particular user, and if it wasn't it looped, etc. with a manual timeout. This was done because the checks couldn't access the user variable. In that situation, you would now instead use get_msg_check(reaction_user=user), which returns a check function for the inputed user.)