Timeout now resets on turn. Update help
Timeout can now be changed using the timeout variable. Also, when the pages are "turned", the timeout resets, which makes more sense (users that are currently turning pages shouldn't have the message suddenly deleted while they're using it.) It also makes more sense with custom messages (when a message is inputed in `msg`, it might not necessarily have a `delete_after` value.

Help string was updated with the new variables recently added, and the help for editable_content is now clearer