Create globals to hold placeholder patterns
Asked in review. There is now a global arg and patterns tuple. Regex patterns are now already compiled globally instead of being recompiled everytime. There is now a patterns_and_values dict property in pstrings that is iterated over so that there is less code repetition and so that its now easier to add new placeholders.