Add p-strings support, small fixes
Replace most of the on_message logic to a PStringsEncoding,parser. Users can now use the placeholders such as %user%, %channels%, %1%, %[]%. Added a help section to know how to use placeholders. number_check now checks that the number is sent in the right channel. max_concurrency is now set to be max 1 per user instead of per channel. The custom reacts and proposed custom reacts lists now have a max length to prevent the embed from not being sent