Improve req/opt vars logic, comments fix
Change the logic of convert. A copy of the input dictionnary is now made, from which keys are removed when they are filled. If there are still keys in this at the end, check if they are required (and raise error) or optional (and fill them with default value).

Corrections to comments, and add attribution