Fix foreign key for currency, move functions
DB foreign key is now activated before transactions, and members are added to the Members table before the transaction if they aren't already in it. The add_member_if_needed function from score.py is moved to an util, since it is now also used by currency and could be used by other functions in the future. The on_user_update listener from score.py that changes usernames on the Members table after users update them is moved to Main.py as it is not directly related to score.py. A on_member_join listener to add users to the Members table when they join is also added to Main.py