Rewrite score completely: Log all reactions and get advanced stats for all emojis (#319)
* Add Reactions table, remove Score from Members

**Why this change was necessary**
The Reactions table is to get better stats on reactions. Score is now computed from Reactions and thus not needed anymore in Members. DisplayName is renamed Name to better reflect how it is now used (See below)

**What this change does**
Adds a Reactions table that, for every reaction, will keep the ID of the person that reacted (ReacterID),the ID of the person that wrote the message (ReacteeID), the name of the reaction (this is the result of str(emoji), that is, either the emoji directly for unicode directly, or <:emojiname:emojiid> for custom emoji), and the ID of the message. The ReacterID and ReacteeID reference IDs in the Members table. Each row is unique since a certain user can only react a certain emoji on a certain message once.

Since score is now calculated directly from that table, there is no need to cache score in the Members table anymore. DisplayName is renamed Name to avoid confusion since it will now use the username#descriminator instead (or if that is unavaible, the ID). The reference to DisplayName in currency.py is changed to Name.

* Rewrite score.py to work with all emojis

**Why this change was necessary**
This change was necessary to get better stats, and be able to track reactions better.

**What this change does**
-This is a near complete rewrite of score.py
-On raw reaction add/remove (for every emoji), the ReacterID, ReacteeID, ReactionName and MessageID is added/removed to/from the Reactions table
-If the Reacter or Reactee users aren't in the Members table, they are added. If the user isn't in the server anymore, take a little bit more time to try and find their name in the rest of Discord (using fetch_user instead of get_user). If the account is deleted, use the ID as name.
-If a user updates their username, update it in the database (this should possibly be moved out of score.py in the future)
-Rewrite ?score: It now calculates score directly from the Reactions table (the result is still nearly instant). If no argument is given, gives the total score for the user who called it (upmartlets - downmartlets). If a mention is given, gives the total score for the user mentioned. If an emoji is given (but no mention), gives the score for that emoji for the user who called it. If an emoji and a mention is given (*in any order*), gives the score for that emoji for the user mentioned. (For all of these, self-reacts or ignored).
-Rewrite ?ranking: Also calculates scores directly from the Reactions table (and even tho it must calculate score for many users, the result is still nearly instant!). If no argument is given, the total score ranking is given (Users that never got an upmartlet or downmartlet won't appear there). If an emoji is given, gives the ranking for that emoji. Now uses an embed instead of the fancy_grid.

* Code style changes

**Why this change was necessary**
Asked in reviews

**What this change does**
-Remove explicit mention of .keys when iterating over a dictionnary
-Use chain instead of adding up last of keys
-if not ... instead of if len(...) == 0

* Code style, handle rare case

Change unnecessary try except to if statement (asked in review)
Handle a case where upvotes were never used on the server

* YAPF

* Format schema, use primary key id in Members

* Create emoji.json

* New complete rewrite of score.py

Rewritten to use three commands, ?score, ?ranking and ?emojiranking, which can take many arguments, see the help strings for the three functions. These arguments can give all of the stats that were requested in the past, for examples which user upvotes you the most, which reactions are the most used in the server, who reacted the most with a certain emoji, etc.

The sqlite statements have been rewritten to include these arguments but still keep only one or two statement by function (those statements get safely edited with the additional arguments) . They have also been rewritten to include additional tasks that were done by python loops/maps before, such as counting how many times each person used a reaction, numbering the entries, adding "tries" after every number, etc.

The code that was used to make the initial Members and Reactions table has been included in a comment at the bottom. See that comment for more info.

* Yapf

* Small fixes

Emojiname flag is now only added in the converters_dict when the emoji param is True.

A sum to know if more than one element has been specified was broken and is now fixed

Replaced an if by an elif

* yapf wants this....

ok yapf...

* Fix foreign key for currency, move functions

DB foreign key is now activated before transactions, and members are added to the Members table before the transaction if they aren't already in it. The add_member_if_needed function from score.py is moved to an util, since it is now also used by currency and could be used by other functions in the future. The on_user_update listener from score.py that changes usernames on the Members table after users update them is moved to Main.py as it is not directly related to score.py. A on_member_join listener to add users to the Members table when they join is also added to Main.py

* yapf

* Style changes

Requested in review

* Create ThirdPartyNotices.md

* Update table building functions comment

time.clock() is removed in python 3.8. Add_member_if_needed now uses a cursor as argument