Add gif support to banner contest (#503)
Add gif support to banner contest. Output error if user sends image bigger than 10 mb (note that the image much actually be a little bit smaller than that to account for the overlays, so there could be a problem in that edge case). Moved some repetitive code inside of a single function